Move StudyWindow to separate module
