+++ setup.cfg
@@ -18,6 +18,7 @@ install_requires =
 packages =
     language_practice
     language_practice.web
+    language_practice.gui
 
 scripts =
     language-practice
